+++ frontend/src/App.jsx
@@ -3,13 +3,10 @@ import React from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 // import { AuthProvider } from './components/AuthContext/AuthContext';
 import HomeComponent from './components/homeComponent/HomeComponent';
-
-import Panel from './components/Panel/Panel';
-import Docentes from './components/panels/docentes';
-import Alumnos from './components/panels/alumnos';
-// import FormAdmin from './components/formLogin/LoginAdmin';
-// import OpcionesComponent from './components/opciones/Opciones';
-
+// import Admins from './components/adminComponent/Admins';
+import Panelg from './components/Panel/PanelAdmin/Panel';
+import Docentes from './components/Panel/PanelDocente/docentes';
+import Alumnos from './components/Panel/PanelAlumno/alumnos';
 import { AuthProvider } from './components/AuthContext/AuthContext'; 
 // import Panel from './components/Panel/Panel';
 
@@ -22,7 +19,7 @@ function App() {
       <BrowserRouter>
         <Routes>
           <Route path="/" element={<HomeComponent />} />
-          <Route path="/panel" element={<Panel />} />
+          <Route path="/panel" element={<Panelg/>} />
           <Route path="/panel-docente" element={<Docentes />} />
           <Route path="/panel-alumno" element={<Alumnos />} />
           {/* <Route path="/loginadmin" element={<FormAdmin />} />
