+++ frontend/src/components/Panel/PanelAdmin/cursos/cursos.jsx
