--- frontend/src/components/Panel/cursos/cursos.module.css
@@ -1,172 +1,0 @@
-.section {
-  padding: 20px;
-  background-color: white;
-  border-radius: 8px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-  margin: 20px;
-}
-
-.addButtonContainer {
-  margin-bottom: 20px;
-  text-align: right;
-  padding: 0 20px;
-}
-
-.addButton {
-  background-color: #4CAF50;
-  color: white;
-  padding: 10px 20px;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-  font-size: 16px;
-}
-
-.addButton:hover {
-  background-color: #45a049;
-}
-
-.filterContainer {
-  display: flex;
-  gap: 20px;
-  margin-bottom: 20px;
-}
-
-.searchContainer {
-  flex: 1;
-}
-
-.searchInput {
-  width: 100%;
-  padding: 8px;
-  border: 1px solid #ddd;
-  border-radius: 4px;
-  font-size: 16px;
-}
-
-.selectContainer {
-  min-width: 200px;
-}
-
-.cycleSelect {
-  width: 100%;
-  padding: 8px;
-  border: 1px solid #ddd;
-  border-radius: 4px;
-  font-size: 16px;
-}
-
-.table {
-  width: 100%;
-  border-collapse: collapse;
-  margin-top: 20px;
-}
-
-.table th,
-.table td {
-  padding: 12px;
-  text-align: left;
-  border-bottom: 1px solid #ddd;
-}
-
-.table th {
-  background-color: #f5f5f5;
-  font-weight: bold;
-}
-
-.tableRow {
-  cursor: pointer;
-  transition: background-color 0.2s;
-}
-
-.tableRow:hover {
-  background-color: #f0f0f0;
-}
-
-.form {
-  max-width: 500px;
-  margin: 0 auto;
-}
-
-.formGroup {
-  margin-bottom: 20px;
-}
-
-.formGroup label {
-  display: block;
-  margin-bottom: 5px;
-  font-weight: bold;
-}
-
-.formInput {
-  width: 100%;
-  padding: 8px;
-  border: 1px solid #ddd;
-  border-radius: 4px;
-  font-size: 16px;
-}
-
-.formButtons {
-  display: flex;
-  gap: 10px;
-  justify-content: flex-end;
-}
-
-.submitButton,
-.cancelButton {
-  padding: 10px 20px;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-  font-size: 16px;
-}
-
-.submitButton {
-  background-color: #4CAF50;
-  color: white;
-}
-
-.submitButton:hover {
-  background-color: #45a049;
-}
-
-.submitButton:disabled {
-  background-color: #cccccc;
-  cursor: not-allowed;
-}
-
-.cancelButton {
-  background-color: #f44336;
-  color: white;
-}
-
-.cancelButton:hover {
-  background-color: #da190b;
-}
-
-.header {
-  display: flex;
-  align-items: center;
-  gap: 20px;
-  margin-bottom: 20px;
-}
-
-.backButton {
-  background-color: #666;
-  color: white;
-  border: none;
-  padding: 8px 16px;
-  border-radius: 4px;
-  cursor: pointer;
-  font-size: 16px;
-}
-
-.backButton:hover {
-  background-color: #555;
-}
-
-.noCourses {
-  text-align: center;
-  color: #666;
-  margin-top: 20px;
-}
