+++ frontend/vite.config.js
@@ -4,4 +4,8 @@ import react from '@vitejs/plugin-react'
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [react()],
+  server: {
+    host: true, // Permite acceder desde tu red local
+    port: 5173, // Cambia el puerto si es necesario
+  },
 })
